django/traditional_-culture_-graduation-0b78220147f1928b7f1ae49bac4c4b1b1e639019/BackEnd/traditionalSite/requirements.txt
django
djangorestframework
django-cors-headers
pymysql
requests
